# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NADITOS API
  description: |
    National Digital Transport Operating System — public HTTP API.

    Every request flows through the API gateway, which:
      - injects an `X-Request-Id` header,
      - verifies the bearer JWT for protected routes,
      - enforces per-route role gates and per-(tenant, route) rate limits.

    Tenant identity is carried in the JWT claim `tid`. Clients may send
    `X-Tenant-Id` for routes that don't require auth (login); for
    authenticated routes the gateway rejects mismatches.
  version: 1.0.0
servers:
  - url: https://api.naditos.example
    description: Production gateway
  - url: http://localhost:8080
    description: Local docker-compose
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    EvidenceIn:
      type: object
      required: [kind, s3_key, sha256, bytes, taken_at]
      properties:
        kind:    { type: string, enum: [photo, video, signature, document] }
        s3_key:  { type: string }
        sha256:  { type: string }
        bytes:   { type: integer, format: int64 }
        taken_at: { type: string, format: date-time }
    FineIssueRequest:
      type: object
      required: [plate, offence_code, evidence, geo_lat, geo_lng, device_id]
      properties:
        plate:           { type: string }
        offence_code:    { type: string }
        driver_license:  { type: string, nullable: true }
        geo_lat:         { type: number, format: double }
        geo_lng:         { type: number, format: double }
        geo_accuracy_m:  { type: number, format: float }
        device_id:       { type: string }
        notes:           { type: string }
        evidence:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/EvidenceIn" }
    Fine:
      type: object
      properties:
        id:           { type: string, format: uuid }
        plate:        { type: string }
        offence_code: { type: string }
        amount:       { type: string, description: "decimal as string" }
        currency:     { type: string }
        status:
          type: string
          enum: [issued, warned, overdue, paid, disputed, escalated, seized, court, cancelled]
        issued_at:    { type: string, format: date-time }
        due_at:       { type: string, format: date-time }
        issued_by:    { type: string, format: uuid }
...
security:
  - bearerAuth: []
paths:
...
  /v1/fines:
    get:
      summary: List recent fines (tenant-scoped)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Fine" }
    post:
      summary: Issue a fine — anti-corruption gates apply
      description: |
        Officer cannot set the amount — server reads
        `regulation_offences.base_amount` for the active country pack.
        Evidence (>=1 photo with sha256, taken_at) is REQUIRED.
        Duplicates within the offence's `duplicate_window_min` for the
        same `vehicle_id × offence_code` are rejected.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/FineIssueRequest" } } }
      responses:
        '201': { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/Fine" } } } }
        '400': { description: missing evidence, unknown offence, etc. }
        '409': { description: duplicate fine within window }
...
  /v1/fines/{id}/pay:
    post:
      summary: Initiate payment via the configured provider
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: payment intent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  intent_id:     { type: string }
                  client_secret: { type: string }
                  status:        { type: string }
...